cicd: add missing permissions
When you call a reusable workflow (workflow_call) that requires permissions, GitHub mandates that you explicitly grant those permissions from the caller, or the job fails.
This is to prevent a situation where someone could submit a malicious PR that runs code with write access to issues or PRs.

Addresses: #813

Signed-off-by: Dan Calavrezo <dan.calavrezo.ext@qorix.ai>